modified Auth Token interface; condensed create & verify auth tokens
