--- app/(api)/_datalib/auth/createToken.ts
@@ -1,8 +1,0 @@
-import { auth_expiration } from '@apidata/configs';
-import jwt, { Secret } from 'jsonwebtoken';
-
-export async function createAuthToken(data: object) {
-  return jwt.sign(data, process.env.JWT_SECRET as Secret, {
-    expiresIn: `${auth_expiration}h`,
-  });
-}
