--- app/(api)/_types/User.ts
@@ -1,11 +1,0 @@
-interface UserInt {
-  _id: string;
-  email: string;
-  password: string;
-
-  // optional fields
-  // username?: string;
-  // date_created?: Date;
-}
-
-export default UserInt;
